fix: review comments

+++ apps/api/src/app/sync-batch/services/sync-batch.validation.ts
@@ -3,7 +3,7 @@ import { DEFAULT_SYNC_BATCH_SINCE_DAYS_AGO } from "./sync-batch.service";
 
 export const scheduleSyncBatchValidationSchema = z.object({
   workspaceId: z.number(),
-  sinceDaysAgo: z.number().min(1).max(DEFAULT_SYNC_BATCH_SINCE_DAYS_AGO),
+  sinceDaysAgo: z.number().min(1).max(DEFAULT_SYNC_BATCH_SINCE_DAYS_AGO).int(),
 });
 
 export type ScheduleSyncBatchInput = z.infer<
